[NCL-9807] Fix (#2) starting of extraction rebuildMode into build metadata
