+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/ExpliciteDependenciesTest.java
@@ -17,9 +17,9 @@
  */
 package org.jboss.pnc.coordinator.test;
 
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.spi.BuildOptions;
-import org.jboss.pnc.enums.RebuildMode;
 import org.junit.Before;
 import org.junit.Test;
