+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/InGroupDependentBuildsTest.java
@@ -17,11 +17,11 @@
  */
 package org.jboss.pnc.coordinator.test;
 
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.common.json.ConfigurationParseException;
 import org.jboss.pnc.mock.repository.BuildConfigurationRepositoryMock;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationSet;
-import org.jboss.pnc.enums.RebuildMode;
 import org.jboss.pnc.spi.datastore.DatastoreException;
 import org.jboss.pnc.spi.exception.CoreException;
 import org.junit.Before;
