+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/SkippingBuiltConfigsTest.java
@@ -17,9 +17,9 @@
  */
 package org.jboss.pnc.coordinator.test;
 
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.common.json.ConfigurationParseException;
 import org.jboss.pnc.enums.BuildStatus;
-import org.jboss.pnc.enums.RebuildMode;
 import org.jboss.pnc.mock.repository.BuildConfigurationRepositoryMock;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationSet;
