+++ integration-test-rex/src/test/java/org/jboss/pnc/integrationrex/DependentBuildsTest.java
@@ -21,6 +21,7 @@ import com.github.tomakehurst.wiremock.WireMockServer;
 import lombok.extern.slf4j.Slf4j;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.auth.KeycloakClient;
 import org.jboss.pnc.client.BuildClient;
 import org.jboss.pnc.client.BuildConfigurationClient;
@@ -34,7 +35,6 @@ import org.jboss.pnc.dto.GroupBuild;
 import org.jboss.pnc.dto.GroupConfiguration;
 import org.jboss.pnc.dto.requests.GroupBuildRequest;
 import org.jboss.pnc.enums.BuildStatus;
-import org.jboss.pnc.enums.RebuildMode;
 import org.jboss.pnc.integrationrex.mock.BPMResultsMock;
 import org.jboss.pnc.integrationrex.utils.BuildUtils;
 import org.jboss.pnc.integrationrex.utils.ResponseUtils;
