+++ remote-build-coordinator/src/test/java/org/jboss/pnc/remotecoordinator/test/dependencies/InGroupDependentBuildsTest.java
@@ -17,9 +17,9 @@
  */
 package org.jboss.pnc.remotecoordinator.test.dependencies;
 
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.common.graph.GraphStructureException;
 import org.jboss.pnc.common.json.ConfigurationParseException;
-import org.jboss.pnc.enums.RebuildMode;
 import org.jboss.pnc.mock.repository.BuildConfigurationRepositoryMock;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationSet;
