+++ remote-build-coordinator/src/test/java/org/jboss/pnc/remotecoordinator/test/event/StatusUpdatesIT.java
@@ -20,11 +20,11 @@ package org.jboss.pnc.remotecoordinator.test.event;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.junit.InSequence;
+import org.jboss.pnc.api.enums.RebuildMode;
 import org.jboss.pnc.api.enums.orch.CompletionStatus;
 import org.jboss.pnc.common.util.ObjectWrapper;
 import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.enums.BuildStatus;
-import org.jboss.pnc.enums.RebuildMode;
 import org.jboss.pnc.mock.model.builders.TestProjectConfigurationBuilder;
 import org.jboss.pnc.mock.spi.RepositoryManagerResultMock;
 import org.jboss.pnc.model.BuildConfigSetRecord;
